+++ getX.h
@@ -1,6 +1,6 @@
 #ifndef getX
 
 
-void get_x(double** input);
+void get_x(float** input);
 
 #endif
